requirements.txt
Django>=4.0
djangorestframework
drf-yasg
celery
redis
django-celery-results
pandas
XlsxWriter
